Moved hook to correct folder
